requirements.txt
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.0
langchain==0.2.6
langchain_chroma==0.1.2
langchain_community==0.2.6
langchain_core==0.2.11
langchain_google_community==1.0.6
langchain_groq==0.1.6
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.2
python-dotenv==1.0.1
streamlit==1.36.0
